--- frontend/src/components/MissionDetailButtonGroup/MissionDetailButtonGroup.styled.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-export const MissionDetailButtonGroupContainer = styled.div`
-  display: flex;
-  gap: 2rem;
-  justify-content: flex-end;
-`;
-
-export const Button = styled.button`
-  background-color: var(--grey-300);
-  font-size: 1.5rem;
-  padding: 0.75rem;
-  border-radius: 0.5rem;
-  color: var(--black-color);
-`;
