--- frontend/src/components/MissionDetailContent/MissionDetailContent.styled.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const MissionDescription = styled.div`
-  padding: 0 2.4rem;
-  font-size: 1.8rem;
-
-  background-color: var(--grey-50);
-  border-radius: 1rem;
-  padding: 2rem;
-`;
